requirements.txt
amqp==2.2.2
appnope==0.1.0
awsebcli==3.12.1
billiard==3.5.0.3
blessed==1.14.2
botocore==1.7.45
celery==4.1.0
cement==2.8.2
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.7
coolsms-python-sdk==2.0.3
decorator==4.1.2
Django==1.11.7
django-cors-headers==2.1.0
django-extensions==1.9.7
djangorestframework==3.7.3
docker-py==1.7.2
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
google-api-python-client==1.6.4
httplib2==0.10.3
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
jmespath==0.9.3
kombu==4.1.0
oauth2client==4.1.2
olefile==0.44
packaging==16.8
parso==0.1.1
pathspec==0.5.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
pip-review==1.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.9.1
rsa==3.4.2
selenium==3.7.0
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
websocket-client==0.44.0
